pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
exclude = [
    '/monkey_patch\.py$',
]
follow_imports = "normal"
no_implicit_reexport = true
plugins = "pydantic.mypy"
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.poetry]
authors = ["JOJ Dev Team <liuyh615@126.com>"]
description = ""
documentation = "https://joint-online-judge.github.io/"
homepage = "https://joj.sjtu.edu.cn/"
keywords = ["oj"]
license = "MIT"
name = "joj-horse"
packages = [
    { include = "joj" },
]
readme = "README.md"
repository = "https://github.com/joint-online-judge/horse/"
version = "0.1.0"

[tool.poetry.dependencies]
aiocache = {extras = ["redis", "msgpack"], version = "^0.11.1"}
aioredis = {version = "^1.3.1"}
aioredlock = "^0.7.2"
alembic = "^1.7.4"
asgi-lifespan = {version = "^1.0.1", optional = true}
asyncpg = "^0.24.0"
asynctest = {version = "^0.13.0", optional = true}
boto3 = "^1.18.38"
celery = {extras = ["redis"], version = "^5.1.2"}
click = "7.1.2"
click-option-group = "^0.5.3"
coverage = {version = "^5.5", optional = true}
fastapi = {extras = ["all"], version = "==0.73.0"}
fastapi-jwt-auth = {git = "https://github.com/joint-online-judge/fastapi-jwt-auth.git", branch = "feature/pyjwt-2.1.0"}
fastapi-utils = "^0.2.1"
fastapi-versioning = "^0.10.0"
GitPython = "^3.1.20"
greenletio = "^0.9.0"
hiredis = "^2.0.0"
httpx = "^0.19.0"
httpx-oauth = "^0.3.7"
joj-elephant = {git = "https://github.com/joint-online-judge/elephant.git", branch = "master"}
lakefs-client = "0.47.0"
loguru = "^0.5.3"
makefun = "^1.12.1"
oauth-jaccount = "^0.1.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pbr = "^5.6.0"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
pydantic-universal-settings = "^0.1.5"
PyJWT = "^2.3.0"
pytest = {version = "^6.2.5", optional = true}
pytest-asyncio = {version = "^0.15.1", optional = true}
pytest-cov = {version = "^2.12.1", optional = true}
pytest-depends = {version = "^1.0.1", optional = true}
pytest-lazy-fixture = {version = "^0.6.3", optional = true}
python = "^3.10"
python-multipart = "^0.0.5"
rollbar = "^0.16.2"
sentry-sdk = "^1.3.1"
SQLAlchemy-Utils = "^0.37.8"
sqlmodel = "^0.0.6"
starlette-context = "^0.3.3"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}
pre-commit = "^2.15.0"

[tool.poetry.extras]
test = [
    "asgi-lifespan",
    "asynctest",
    "coverage",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-depends",
    "pytest-lazy-fixture",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.ini_options]
filterwarnings = "ignore::DeprecationWarning"
log_cli = 1
log_cli_level = "INFO"

[tool.semantic_release]
branch = "master"
build_command = "pip install poetry && poetry build"
upload_to_pypi = false
upload_to_release = false
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variable = ["joj/horse/__init__.py:__version__"]
